Rename Option to CheckboxOption
